Move reference-helpers to a better location (utils)
